requirements.txt
# Core dependencies
dbt-snowflake==1.7.1
snowflake-connector-python==3.6.0
great-expectations==0.18.8

# AWS
boto3==1.34.23
awscli==1.32.23

# Data generation
Faker==22.0.0
pandas==2.1.4
pyarrow==14.0.2

# Data quality and validation
sqlfluff==3.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Visualization
graphviz==0.20.1
diagrams==0.23.4

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

